♻️Refactor(groups, roles): 파일명 수정. groups->group, roles->role
